feat: refactor utility functions and add rip7560 signature function
