+++ src/utils/index.ts
@@ -4,24 +4,29 @@ import {
     getRandomAddress
 } from "./constructRip7560Transaction";
 import { constructUserOp } from "./constructErc4337UserOp";
-import { getRlpHash } from "./generateRlpHash";
 import { getCallData } from "./getCallData";
 import { getDeployerData, getAddress } from "./getDeployerData";
 import { getNonce, getNonceFromNonceManager } from "./getNonce";
 import { decodeRevertReason } from "./decodeRevert";
 import { getChainId } from "./getChainId";
+import { 
+    signUserOp,
+    signRip7560Transaction
+} from "./sign";
+
 
 export {
     constructRip7560Transaction,
     constructUserOp,
     getDummyAuthorizationData,
     getRandomAddress,
-    getRlpHash,
     getCallData,
     getDeployerData,
     getAddress,
     getNonce,
     getNonceFromNonceManager,
     decodeRevertReason,
-    getChainId
+    getChainId,
+    signUserOp,
+    signRip7560Transaction
 }
